some light refactoring, add @Authenticated decorator
