--- packages/api/src/business/auth/AuthKeys.ts
@@ -1,3 +1,0 @@
-export enum AuthKeys {
-  Authenticated = "IS_LOGGED_IN",
-}
